Add advantages to sheet
By adding the advantage to the sheet instead of the point total this
prevents violation of Demeters law.

Furthermore, note the inversion of control. We let advantage pay its
cost to points, this allows advantage to encapsulate its data (i.e.
cost) and only the public api of points is used.