Group conceptual build steps together
Some steps form a conceptual group, e.g. SetNameStep, SetTitleStep and
SetReligionStep together set the Identity. This concept is introduced by
applying the Composite Pattern. This treat a group of steps as a single
step.

To see how this is implemented see the abstract class CompositeStep.

For more information on the Composite Pattern see
https://en.wikipedia.org/wiki/Composite_pattern