Explicit handle optional Integer values
The different appearances of the use of a cast or a toString call that
Marcel fixed two commits ago, some of them were there for a reason.

They represented Integer values, which caused a problem because Integers
can not be cast to a String.

I now explicitly handle the integer cases.